feat(dictionary): the word family ends the page
The form tree moves in with the word: every reading's family — the
lemma at the root, its forms by how each hangs off it, climbable both
ways — now ends the word page, after the corpus has said its piece
and above the way out to searching it. The same trees the lemma pages
draw, brought to where the reader already is: the tree came out of
the lemma route into a component the two share, and /dictionary/lemma
still stands for the index. The section's heading waits for the
trees — a reading with nothing hanging off it draws nothing, and a
word whose every reading is bare draws no section at all: an empty
table under a heading would only say the feature exists.
