+++ CorpusSearch/ClientApp/src/components/LemmaTree.css
@@ -1,0 +1,122 @@
+/* the form tree, drawn as one — trunk, branches, leaves (LemmaTree.tsx).
+   Worn by its own page at /dictionary/lemma and, embedded, at the end of a
+   word page's entries. */
+
+/* what the root hangs off, read above it: the graph climbs both ways */
+.dict-lemma-parent {
+    margin: 0;
+    color: var(--c-muted);
+    font-size: 14px;
+}
+
+.dict-lemma-parent-types {
+    font-style: italic;
+    font-size: 13px;
+}
+
+/* the heading is the tree's root: the trunk below hangs straight off it */
+.dict-lemma-root {
+    margin-bottom: 2px;
+    margin-top: 4px;
+}
+
+/* the root inside a word page: heads its tree, not the page */
+.dict-lemma-root-embedded {
+    font-family: Georgia, serif;
+    font-size: 22px;
+    color: #33454d;
+    margin: 0 0 2px;
+}
+
+.dict-lemma-tree,
+.dict-lemma-tree ul {
+    list-style: none;
+    margin: 0;
+    padding: 0;
+    font-family: Georgia, serif;
+}
+
+.dict-lemma-tree {
+    margin: 0 0 6px 7px;
+}
+
+/* the nested list starts under its branch's tick */
+.dict-lemma-tree li > ul {
+    margin-left: 17px;
+}
+
+.dict-lemma-tree li {
+    position: relative;
+    padding: 3px 0 3px 22px;
+}
+
+/* the vertical of the trunk, running the full height of every node but the
+   last, where it stops at the tick */
+.dict-lemma-tree li::before {
+    content: "";
+    position: absolute;
+    top: 0;
+    bottom: 0;
+    left: 0;
+    border-left: 1px solid var(--c-muted);
+}
+
+.dict-lemma-tree li:last-child::before {
+    bottom: auto;
+    height: 14px;
+}
+
+/* the tick from trunk to node */
+.dict-lemma-tree li::after {
+    content: "";
+    position: absolute;
+    top: 14px;
+    left: 0;
+    width: 16px;
+    border-top: 1px solid var(--c-muted);
+}
+
+/* a branch names how its leaves hang off the lemma */
+.dict-lemma-branch {
+    display: inline-block;
+    padding: 1px 6px;
+    border-radius: 2px;
+    background-color: var(--c-border-meta);
+    font-size: 13px;
+    font-weight: bold;
+    letter-spacing: 0.5px;
+}
+
+/* how often the corpus says a node's spelling, as the walk counts uses */
+.dict-lemma-count {
+    color: var(--c-accent);
+    font-size: 12px;
+}
+
+/* the book behind a node no text uses: without it a greyed form reads as a
+   phantom, when in fact Cregeen prints it */
+.dict-lemma-source {
+    color: var(--c-muted);
+    font-size: 12px;
+}
+
+.dict-page-word .dict-lemma-count {
+    font-size: 15px;
+    font-weight: normal;
+}
+
+@media (max-width: 600px) {
+    /* a deep chain (dooinney → deiney → e gheiney → gheiney) must not walk
+       off a phone: shallower steps, same shape */
+    .dict-lemma-tree li {
+        padding-left: 12px;
+    }
+
+    .dict-lemma-tree li > ul {
+        margin-left: 8px;
+    }
+
+    .dict-lemma-tree li::after {
+        width: 8px;
+    }
+}
